+++ src/main/java/roomescape/ReservationController.java
@@ -6,14 +6,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import roomescape.exceptions.EmptyNameException;
-import roomescape.exceptions.PastDateTimeException;
-import roomescape.exceptions.ReservationNotFoundException;
+import roomescape.domain.Reservation;
+import roomescape.domain.Reservations;
 
 @Controller
 public class ReservationController {
@@ -40,15 +38,5 @@ public class ReservationController {
     public ResponseEntity<Void> delete(@PathVariable Long id) {
         reservations.delete(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
-    @ExceptionHandler({EmptyNameException.class, PastDateTimeException.class})
-    public ResponseEntity<String> handleIllegalReservation(Exception exception) {
-        return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body(exception.getMessage());
-    }
-
-    @ExceptionHandler({ReservationNotFoundException.class})
-    public ResponseEntity<String> handleReservationNotFoundException(Exception exception) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(exception.getMessage());
     }
 }
