+++ src/main/java/roomescape/domain/Reservations.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain;
 
 import java.util.ArrayList;
 import java.util.List;
