+++ src/main/java/roomescape/exceptions/EmptyNameException.java
@@ -2,6 +2,6 @@ package roomescape.exceptions;
 
 public class EmptyNameException extends RuntimeException {
     public EmptyNameException() {
-        super("이름이 비어있을 수 없습니다.");
+        super("이름이 비어 있을 수 없습니다.");
     }
 }
